extract standalone mod
